+++ NEWS.md
@@ -1,4 +1,4 @@
-## 2020-11-29 v2.5.0-SNAPSHOT
+## 2020-11-29 v3.0.0-SNAPSHOT
 * [MODSOURMAN-338](https://issues.folio.org/browse/MODSOURMAN-338) Change chunk processing to use kafka
 * [MODSOURCE-177](https://issues.folio.org/browse/MODSOURCE-177) Use kafka to save chunks of parsed records in SRS
 * [MODSOURMAN-380](https://issues.folio.org/browse/MODSOURMAN-380) Expand journalRecord entity with "title" property
