+++ mod-source-record-manager-server/src/main/java/org/folio/services/ChunkProcessingService.java
@@ -3,9 +3,6 @@ package org.folio.services;
 import io.vertx.core.Future;
 import org.folio.dataimport.util.OkapiConnectionParams;
 import org.folio.rest.jaxrs.model.RawRecordsDto;
-import org.folio.rest.jaxrs.model.Record;
-
-import java.util.List;
 
 public interface ChunkProcessingService {
