+++ mod-source-record-manager-server/src/main/java/org/folio/services/RecordsPublishingService.java
@@ -2,7 +2,6 @@ package org.folio.services;
 
 import io.vertx.core.Future;
 import org.folio.dataimport.util.OkapiConnectionParams;
-import org.folio.rest.jaxrs.model.ParsedRecordDto;
 import org.folio.rest.jaxrs.model.Record;
 
 import java.util.List;
