+++ mod-source-record-manager-server/src/main/java/org/folio/verticle/AbstractConsumersVerticle.java
@@ -1,0 +1,74 @@
+package org.folio.verticle;
+
+import io.vertx.core.AbstractVerticle;
+import io.vertx.core.Promise;
+import org.folio.kafka.AsyncRecordHandler;
+import org.folio.kafka.GlobalLoadSensor;
+import org.folio.kafka.KafkaConfig;
+import org.folio.kafka.KafkaConsumerWrapper;
+import org.folio.kafka.KafkaTopicNameHelper;
+import org.folio.kafka.SubscriptionDefinition;
+import org.folio.spring.SpringContextUtil;
+import org.folio.util.pubsub.PubSubClientUtils;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.support.AbstractApplicationContext;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public abstract class AbstractConsumersVerticle extends AbstractVerticle {
+
+  //TODO: get rid of this workaround with global spring context
+  protected static AbstractApplicationContext springGlobalContext;
+
+  private static final GlobalLoadSensor globalLoadSensor = new GlobalLoadSensor();
+
+  @Autowired
+  @Qualifier("newKafkaConfig")
+  private KafkaConfig kafkaConfig;
+
+  @Value("${srm.kafka.DataImportConsumer.loadLimit:5}")
+  private int loadLimit;
+
+  private List<KafkaConsumerWrapper<String, String>> consumerWrappersList = new ArrayList<>();
+
+  @Override
+  public void start(Promise<Void> startPromise) {
+    context.put("springContext", springGlobalContext);
+
+    SpringContextUtil.autowireDependencies(this, context);
+
+    getEvents().forEach(event -> {
+      SubscriptionDefinition subscriptionDefinition = KafkaTopicNameHelper
+        .createSubscriptionDefinition(kafkaConfig.getEnvId(),
+          KafkaTopicNameHelper.getDefaultNameSpace(),
+          event);
+      consumerWrappersList.add(KafkaConsumerWrapper.<String, String>builder()
+        .context(context)
+        .vertx(vertx)
+        .kafkaConfig(kafkaConfig)
+        .loadLimit(loadLimit)
+        .globalLoadSensor(globalLoadSensor)
+        .subscriptionDefinition(subscriptionDefinition)
+        .build());
+    });
+
+    consumerWrappersList.forEach(consumerWrapper -> consumerWrapper
+      .start(getHandler(), PubSubClientUtils.constructModuleName())
+      .onComplete(ar -> startPromise.complete()));
+  }
+
+  @Override
+  public void stop(Promise<Void> stopPromise) {
+    consumerWrappersList.forEach(consumerWrapper -> consumerWrapper
+      .stop()
+      .onComplete(ar -> stopPromise.complete()));
+  }
+
+  public abstract List<String> getEvents();
+
+  public abstract AsyncRecordHandler<String, String> getHandler();
+
+}
