+++ mod-source-record-manager-server/src/main/java/org/folio/verticle/RawMarcChunkConsumersVerticle.java
@@ -1,69 +1,29 @@
 package org.folio.verticle;
 
-import io.vertx.core.AbstractVerticle;
-import io.vertx.core.Promise;
 import org.folio.kafka.AsyncRecordHandler;
-import org.folio.kafka.GlobalLoadSensor;
-import org.folio.kafka.KafkaConfig;
-import org.folio.kafka.KafkaConsumerWrapper;
-import org.folio.kafka.KafkaTopicNameHelper;
-import org.folio.kafka.SubscriptionDefinition;
-import org.folio.spring.SpringContextUtil;
-import org.folio.util.pubsub.PubSubClientUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.support.AbstractApplicationContext;
+
+import java.util.Collections;
+import java.util.List;
 
 import static org.folio.rest.jaxrs.model.DataImportEventTypes.DI_RAW_MARC_BIB_RECORDS_CHUNK_READ;
 
-public class RawMarcChunkConsumersVerticle extends AbstractVerticle {
-
-  //TODO: get rid of this workaround with global spring context
-  private static AbstractApplicationContext springGlobalContext;
-
-  private static final GlobalLoadSensor globalLoadSensor = new GlobalLoadSensor();
+public class RawMarcChunkConsumersVerticle extends AbstractConsumersVerticle {
 
   @Autowired
   @Qualifier("RawMarcChunksKafkaHandler")
   private AsyncRecordHandler<String, String> rawMarcChunksKafkaHandler;
 
-  @Autowired
-  @Qualifier("newKafkaConfig")
-  private KafkaConfig kafkaConfig;
-
-  @Value("${srm.kafka.RawMarcChunkConsumer.loadLimit:5}")
-  private int loadLimit;
-
-  private KafkaConsumerWrapper<String, String> consumerWrapper;
-
   @Override
-  public void start(Promise<Void> startPromise) {
-    context.put("springContext", springGlobalContext);
-
-    SpringContextUtil.autowireDependencies(this, context);
-
-    SubscriptionDefinition subscriptionDefinition = KafkaTopicNameHelper
-      .createSubscriptionDefinition(kafkaConfig.getEnvId(),
-        KafkaTopicNameHelper.getDefaultNameSpace(),
-        DI_RAW_MARC_BIB_RECORDS_CHUNK_READ.value());
-
-    consumerWrapper = KafkaConsumerWrapper.<String, String>builder()
-      .context(context)
-      .vertx(vertx)
-      .kafkaConfig(kafkaConfig)
-      .loadLimit(loadLimit)
-      .globalLoadSensor(globalLoadSensor)
-      .subscriptionDefinition(subscriptionDefinition)
-      .build();
-
-    consumerWrapper.start(rawMarcChunksKafkaHandler, PubSubClientUtils.constructModuleName())
-      .onComplete(ar -> startPromise.complete());
+  public List<String> getEvents() {
+    return Collections.singletonList(DI_RAW_MARC_BIB_RECORDS_CHUNK_READ.value());
   }
 
   @Override
-  public void stop(Promise<Void> stopPromise) {
-    consumerWrapper.stop().onComplete(ar -> stopPromise.complete());
+  public AsyncRecordHandler<String, String> getHandler() {
+    return this.rawMarcChunksKafkaHandler;
   }
 
   //TODO: get rid of this workaround with global spring context
